Hierarchical back navigation on the DC screens
The 500 ms hold is now a back action (GoMainMenu renamed NavigateBack):
channel detail and USB-PD input screens return to the DC overview, and
the overview (and every menu screen) returns to the main menu. The
button tracker's legacy duplicate fire on release is gone - with a
hierarchy it would step up two levels.

Flash budget: formatting every u8 write! argument as u32 removes the
core::fmt u8 Display monomorphization (~110 bytes), and the two-line
CV/CC help section is folded into the AWG section.

+++ Firmware/Rust/benchvolt-pd/src/main.rs
@@ -741,7 +741,7 @@ fn main() -> ! {
                     }
                     arb_runtime::write(chunk);
                     let mut response: String<32> = String::new();
-                    write!(&mut response, "OK:ACK:CH{}\r\n", chunk.channel + 1).ok();
+                    write!(&mut response, "OK:ACK:CH{}\r\n", u32::from(chunk.channel) + 1).ok();
                     queue_usb_response(response.as_bytes());
                 }
                 UsbIntent::ArbStart(start) => {
@@ -1156,7 +1156,7 @@ fn main() -> ! {
                 write!(
                     &mut response,
                     "OK:CH{}_ARB_STARTED_PTS:{}\r\n",
-                    start.channel + 1,
+                    u32::from(start.channel) + 1,
                     start.count
                 )
                 .ok();
